+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   component.
 - Use buttons instead of links in `Doggo.action_bar/1`.
 - Add `toolbar` role to `Doggo.action_bar/1`.
+- Use `section` instead of `article` in `Doggo.modal/1`.
+- Add `dismissable` attribute to `Doggo.modal/1`.
 - Depend on `phoenix_storybook ~> 0.6.0`.
 
 ## [0.4.0] - 2023-12-31